Move BrowserPrompt to react-router-dom package and update dependents
